.claude/skills/sqlite-vectordb/scripts/requirements.txt
sqlite-vec>=0.1.1
sentence-transformers>=2.2.0
numpy>=1.24.0
pyyaml>=6.0
